fix: fix multiple same events + add cairo code + seperate tests, legacy and Cairo1
