+++ __tests__/fixtures.ts
@@ -31,8 +31,6 @@ export const compiledOpenZeppelinAccount = readContract('Account');
 export const compiledErc20 = readContract('ERC20');
 export const compiledErc20Echo = readContract('ERC20-echo');
 export const compiledL1L2 = readContract('l1l2_compiled');
-export const compiledEvent = readContractSierra('event');
-export const compiledEventCasm = readContractSierraCasm('event');
 export const compiledTypeTransformation = readContract('contract');
 export const compiledMulticall = readContract('multicall');
 export const compiledTestDapp = readContract('TestDapp');
